+++ packages/plugin/docs/examples/bin/docs.config.js
@@ -1,7 +1,13 @@
 
 /** @type {import( '../../src/main' ).Config} */
 export default {
-	input  : '../../../docs',
-	output : './build/bin',
+	input    : '../../../docs',
+	output   : './build/bin',
+	twoslash : false,
+
+	vitepress : {
+		ignoreDeadLinks : true,
+		sitemap         : { hostname: 'https://dovenv.pigeonposse.com' },
+	},
 }