+++ packages/config/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dovenv/eslint-config
 
+## 1.6.1
+
+### Patch Changes
+
+- update deps
+
 ## 1.6.0
 
 ### Minor Changes